got the tooltips working!
