+++ client/src/App.js
@@ -36,17 +36,17 @@ class App extends Component {
 
   render() {
     return (
-	   //this lets connect to different components of our site
+      //this lets connect to different components of our site
       <BrowserRouter>
         <div>
           <Switch>
-	    <Route path="/" component={NavBar} exact /> 
-      <Route path="/logging" component={ClassLogging} />
-	    <Route path="/signin" component={Home} />
-      <Route path="/ge" component={GERequirements} />
-      <Route path="/calendar" component={Calendar} />
-      <Route path="/major" component={Major} />
-	    <Route component={Error} />
+            <Route path="/" component={NavBar} exact />
+            <Route path="/logging" component={ClassLogging} />
+            <Route path="/signin" component={Home} />
+            <Route path="/ge" component={GERequirements} />
+            <Route path="/calendar" component={Calendar} />
+            <Route path="/major" component={Major} />
+            <Route component={Error} />
           </Switch>
         </div>
       </BrowserRouter>
